Improve efficiency of evolution tracking
Associating a pokemon and its evolutions was using an O(n^2) algorithm.
For 151 pokemon, that's 22801 iterations! Changing it into O(n)
algorithm. So, the pokemon list population should take less than 1% of
the previous time.

To do this the evolution list had to be converted into a Pokemon list.
This has some cascading code change impact, but those make the code more
readable and efficient too (removes a lot of unnecessary int to Pokemon
lookups).