Add Google Now support when running in battery saver mode
It should be possible to change this in the future (with some code
refactoring) to add the capability to use Google Now without actively
having to start/stop the GoIV service.

Technically any fullscreen viewer can be used to share and analyze
existing screenshot files, but the user will have to update the trainer
level in the app to match their level when the picture was taken.