Fix candy name based Pokemon search
If the pokemon name match isn't good, we can't just pick the pokemon
name that matches the candy name perfectly. If the candy name was
detected perfectly, then we'll always end up picking the base pokemon.

Instead, we use the candy name to limit the list of pokemon names we
have to check for similarity. Also, if the candy name is not detected
perfectly, we have to approximately check only 1/3 of the pokemon names
for similarity with the candy name instead of the whole pokedex.

Once we figure out the base pokemon with the best candy name match, we
go through all the pokemon with that candy name and see how well they
match with the detected pokemon name.