Cache OCR results for parts of the bitmap that don't change often
Most of the regions of the image that need OCR don't change often across
different times the same pokemon is analyzed. Things like name and candy
name are even common across different pokemon. So, we can safely cache
the results of OCRing these images.

To save memory, we cache the results using a hash of the image and not
the entire image.

The CP image changes a lot due to changing background. So, there's not
much of a point to cache it. So, we don't cache OCR results for the CP.

Profiling of this code showed that the caching operation takes about
1-3 milliseconds. Whereas OCR takes from 50-260 milliseconds depending
on the image being OCRed. So, this change definitely bring in a lot of
performance and power saving benefits.