Pokefly.onDestroy: Fix crash on removeView (#135)
`infoShown` does not reflect correctly whether the `infoLayout` and
`arcPointer` were added or not; `infoShownReceived` is closer but is
cleared if `batterySaver`.

In particular, if the main activity hasn't replied yet to the request to
OCR the pokemon, then `infoShown` is true, while `infoShownReceived`
isn't true yet and `infoLayout` and `arcPointer` aren't shown, causing
the crash. This causes at least a race condition that this PR fixes.

This is the crash shown as `#3` and `#33` in Crashlytics.

https://www.fabric.io/goiv/android/apps/com.kamron.pogoiv/issues/57bc2435ffcdc042501dbfa7
https://www.fabric.io/goiv/android/apps/com.kamron.pogoiv/issues/57b57b39ffcdc04250dc0403

Now, not only we have a race condition here, but sometimes the main
activity never replies, making this bug *much* easier to hit. Next
commit fixes that, too.