Fix candy bitmap size to avoid errors
The current bitmap was too wide and went past the "info card" of the
pokemon. This extra bit has a random color and gets past replaceColors()
and results in OCR detecting it as "I". This causes problems when trying
to remove the word "CANDY" when it comes in the end. Instead of removing
"CANDY", we end up removing the "I". So, adjust the size to avoid going
past the "info card".