Delete colon and replace hardcoded string
I deleted the colon in "Level: x" because in some languages (like French), there's a space before a colon. The simplest way to deal with that fact seemed to delete the colon and have "Level x".

I also replaced the "AVG" hardcoded string by the new AVG string that can be translated.