Remove deadlock we had
Here's a potentail deadlock: If scanning fails we produce and send an
empty intent. Right now if we receive that, we set `receivedInfo` even
though we received no info, so the next button press will be ignored.
Usually the flag would be cleared by closing the input/result
dialogs—but since we're showing none, that's not possible!

IIRC I introduced this, sorry. But finally I found it (I think).