Ignore random events due to old screenshots
Sharing any image to other apps seems to trigger ContentObserver
onChange events. So, when an old screenshot of the game is shared to a
random app, it causes GoIV to scan the old screenshot and display the
info dialog. So, add a last modified time check on the file to ignore
these events.