Simplify Nidoran correction
We can just check the last character of the OCR text to identify the
proper gender of the Nidoran. We don't need to do a color check. With
this change, we can apply the Nidoran correction to the candy name
before we search for the final pokemon. This will be useful for future
improvements to Pokemon detection.