Refactor populate advanced result menu & add good default
Refactor: broke down populateEstimateBox into several different methods

Feature:Automatically set estimate pokemon to evolution of scanned poke (good default)

When scanning, in the evolution and powerup estimation box, the evolved
form of the pokemon will now automatically be selected by default, as
that is the most likely selection the user wants.

For example, if scanning a caterpie, metapod will be automatically
selected. If the user scans a pokemon that has no evolution, the pokemon
itself will be selected.

This change fixes #223