Adds bottom margin to manual pokemon select
Moves around code lines as suggested by reformat code in android studio,
and adds a margin to the text input so that it's not covered by the
keyboard when you type.

The only new thing in this commit is really the margin.

Visually this is not great nor optimal, but it makes things easier.