Add PokeSpam calculation
If enabled, compute how many pokemons of the current species can be
evolved with the current amount of candies.

Squashed version of #457 until
7dedc5ed0977e5a5cc0ed676f340a44e1dbc6c80 (excluding the plugin system).

This is just the changes between that commit and
196b638333fb2586ffef60a7a4af91525ed5e504 (the latest commit on master
reflected on that PR).

Open issues: refactor fetching of settings as suggested in the last
comments of the PR.