Cleanup README badges and links
First of two things changed is the cache on the badges was 30 days, so they weren't updating even on hard webpage refresh, so I changed the cache to 6 hours for all but the license one.

Second thing is that using the badges as links isn't obvious enough for users visiting this page just to get the latest download and see the changes. The wording is also awkward between what the badge is trying to communicate and what we're trying to say the link we're using the badge for is actually leading to. For example, the wording of what the download number actually was for (total downloads ever), versus the wording that made it seem like just downloads for the current version. Anyways, the result is I moved the badges all up all in-line so we're not relying on them to be the main links and made user-friendly links to the two thinks casual users need most.