Fixes candy scan
Fixes the candy scan for different pokemon types.

There was an issue with that different pokemon types (water, rock etc)
has different colors on their candy, which meant that the replaceColor
method did not clean out all candy types. The range was VERY high for
some reason (200). This decreases the range, making the text still
readable, but most other colors erased.

It also increases the width of the scan, becase some people indicated
that they might have problems with the previous scan area.,