Further refine the Input Appraisal checkbox formatting
Based on branch justMakeThemCheckboxes, this update further refines the layout, formatting and text strings of the input appraisal dialog input.

- Switched from "LinearLayout (vertical)" for the 4 columns, to "LinearLayout (horizontal)" for 3 rows.
- Set the layout_weight of each CheckBox to "1" to equally divide each row into 4 or 3 columns as necessary.
- Set the textScaleX to .9 on the range and stat rows to slightly reduce the width of each character.
- Re-added longer strings to more similarly match previous spinner wording.

TODO: Tried forcing each Checkbox to "not" wrap the text, but to simply allowed to be clipped off the right-side, but was unable to find a setting that would perform this as expected.  The current implementation tries to limit word length/characters for 2 lines to prevent anything being wrapped to the next line.