Fix a few auto appraisal phrases for more accuracy
mystic_ivrange2_phrase2 used the word "stats" which shows up multiple
 times in other phrases.  Changed to "must" to improve accuracy.

mystic_ivrange4_phrase2 used "opinion" but I found not reference to
that word.  Changed to "estimation".

There are 2 other phrases in which no appropriate value can be
selected as a 2nd phrase.  Updating the values to align with the
technique and comment style from values-de/appraisals.xml