Refines logic for matching appraisal phrases and updates Valor/Instinct
Instinct appraisal phrases use the word "best" quite often.  Due to
this, addInfoFromAppraisalText is updated to only try to match
IVRange when appraisalRangeGroup is not already checked. Also change
"best" to "best of" in instinct_percentage1_phrase2.

In addition, some match-words for Valor are subsets of other words
(strong/strongest, great/greatness).  This change addes a space after
strong and great so as to prevent matching strongest/greatness.
- Has not actually been tested against a live Valor account yet.