Enhance AutoAppraisal sequencing and user experience
This update aims to provide a refined sequence of events and user
interface cues throughout the AutoAppraisal process.

Pokefly.onTouch logic order swap
    This change swaps the logical order of events when a user touches
the screen.  The new order is for the app to first check if the
Appraisal dialog window is VISIBLE, and if so, go straight to
the AutoAppraisal.screenTouched() method, bypassing the normal
Pokefly screenScanHandler.  But, if the appraisal dialog is not
visible, proceed to Pokefly screenScanHandler without engaging
AutoAppraisal code.

AutoAppraisal Rescan if phrase doesn't match.
    This change implements an automatic rescan of the screen if the
AutoAppraisal process is not able to match the scanned phrase.  If
PoGo is still animating the text when the first scan is performed,
this feature allows the AutoAppraisal to rescan up to 3 times at 50ms
intervals to see if the full phrase is eventually able to be seen.

AutoAppraisal visual highlights during appraisal process
    This change adds a rounded rectangle to the Appraisal Dialog box to
highlight the current set of checkboxes that GoIV is currently
looking to populate.  The intent is for this to serve as a visual cue
that GoIV is doing something and to provide some context if/when a
phrase does not scan correctly.
    As part of this change, the flow of "addInfoFromAppraiseText" has
been updated to more accurately track which sections have been
completed and where we should be focusing the appraise phrase matching.

Refactoring of RadioGroup/Checkboxes and related methods/variables
    This change aligns the naming of the variables, methods and
elements related to the Appraisal RadioGroups and Checkboxes.