Fix getBestGuessForEvolutionLine to correctly identify using Candy Name
Prior to this update, the getBestGuessForEvolutionLine method assumed
that only "basePokemon" are used for the Candy names within PoGo.
With the addition of Gen2 Pokemon, specifically the babies, Niantic
has so far decided to keep the Candy Names set to their Gen1
evolutions.  This means that we must now track exactly which Pokemon
name is used by Niantic for Candy naming.

A new integer-array has been added to track this relationship.
Similar to the devolutionNumber, this array maps which Pokemon name
is used for the Candy Name for that evolution line. For the vast
majority, they will point back to themselves (ie: Bulbasaur = 0).
However, for cases like the babies, they will point to the actual
Pokemon in their evolution line that the candy is named after. (ie:
Pichu = 24 (which is Pikachu's array index number)).