Fix IV alignment and sizing of IV Button
The IV Button shape has been changed from oval to rectangle (with
large corner radius) to fix a few issues with alignment and sizing.

There are now 2 iv_button files.  iv_button.xml is the original file
and still contains all layers including the I and V shapes.
The "iv_button_background.xml" has been created without the I and V
so that it can be used as the background for the Pokemon & IV
information.

In addition, the dpToPx method has been updated to (according to
recent online posts and documentation) better represent the true
density of a phone's display.  The placement of the button was also
slightly adjusted to match the apparent alignment of the PoGo menu
button.