Updates to the IV Button and Preview code
Numerous updates and tweaks to the IV Button, it's placement and how
it is managed throughout the scanning/preview process.

Tweaks made to the IV Preview button colors.  The text colors are
using the lighter color scheme, while the shaded colors for the left
and right ends will use more bold colorings for easy identification.

setBackgroundGradient is still under construction.  The idea is to
shade just the far left and far right sides of the pill-shaped quick
preview button.