Improve the CP value detection
Sometimes the CP acronym get scanned as one character instead of two.
This can can lead to the trimming of the first digit of the 'mon CP
value since the code always strips the first two starting characters
from the OCR'd text in order to remove the CP acronym.

This commit introduces the ability to detect the container rectangle of
each character inside the CP cropped image. After that, it ignores all
the characters whose height is lower than the average.
This method effectively detects the CP acronym, regardless of the user
locale and it being suffix or postfix to the value, since it's always
written in a smaller font size.

Resolves: #708