Improve ClipboardModifierActivity UI and UX
Add drag'n'drop to preview list so the user can rearrange its tokens
and add swipe up or down gesture to delete a token.

Increase tokens size in showcase grid to simplify selection.

The user is again able to delete all tokens.

Add a Dialog to tell the user that has unsaved changes on exit.