Changes CPM from float to double
Should make no difference in any calculation, but according to #514 , POGO stores the values as floats, which means that we should also store them as floats to get the exact same values. Since the error is extemely tiny, this will most likely never be an issue, but here's a fix anyways.