Improve text detection with wider Tesseract whitelist
Adding some punctuation symbols (dot and exclamation mark) to its whitelist helps Tesseract to not
misinterpret them with other whitelisted characters.
This is particularly important because Tesseract does not ignore non-whitelisted characters but tries
to interpret them as another whitelisted character, sometimes virtually merging a letter a punctuation
symbol to obtain another whitelisted character.
This could lead to errors especially when detecting appraisal texts that often ends with an exclamation
mark, wich isn't whitelisted: since the last word is often used as key to detect the phrase and, if
Tesseract merges the last character of that keyword with the punctuation, the detection fails.
Adding the dot symbol also helps the appraisal text scan, but also make it easier to detect Mr. Mime.