Removes normalizedPokemonNameMap that took 2 seconds to create
Every time someone pressed the IV button, the inputfraction created a new normalizedPokemonMap - which took 2 seconds on my samsung s7. And it's only used in the very rare case that the user wants to manually type in the pokemon name in the dropdown. This changes to an alternate implementation where it uses levenshteinDistance to just pick whatever pokemon name is closest instead. Functionally the same as far as the user can tell, but takes 2 seconds less time to load.