Auto names generated apk files so they dont have to be manually adjusted
Since the github auto update logic checks for if the name contains "offline" when choosing which apk to download, the apks need to follow a specific naming scheme.  This automates the naming scheme. We need to continue manually changing the 'O' in 'offline' to uppercase for a couple of patches to ensure that old versions continue updating normally.