Extract base data out of Pokemon class
The `Pokemon` class had actually two (or three) roles, which made it
complicated to be used easily.

The one role is obvious and stores the Pokemon's data, like base stats
or number. This is (basically) what the current class still does.

The other role was that it stores which forms of a Pokemon exists, for
example if there are Alolan forms. And it also shared the same objective
as the first role, by being the "normal" Pokemon.

And the last role was that of being a specific form of a Pokemon (like
Alolan). It also had base stats, but it shared the evolutions with the
"base" Pokemon. There was also a "normal" Pokemon instance, which
actually shared the data of the "base" Pokemon mentioned as the second
role. Even more confusing, while the evolutions were shared between all
of these Pokemon, the forms were not. So if you have a Pokemon instance
of this type, you wouldn't actually know that it is a "form" of a
Pokemon. The only way would be to "get" the base Pokemon and see it
there.

(Note from the editor: At least this is how I think it works)

tl;dr: There is a `PokemonBase` class which stores basic data for every
Pokedex number (like it's number and devolution number) and it stores
which forms of Pokemons there are at that Pokedex number (role 2
basically). The old `Pokemon` class has then the specific data for each
Pokemon (like base stats) and also a reference to the base Pokemon for
the rest of the data.