Properly get form for mixed form evolutions
For Pokemons where only certain evolutions use multiple forms, the
current system got confused, because the evolution with the multiple
fomrs didn't have a "devolved" counterpart. So for example alola
Marowaks weren't considered part of the Cubone evolution.

To fix this, this patch does two things:

* Add `getForm(Pokemon)` which properly detects the correct form.
* Add `getEvolutionForms(Pokemon)` which returns all forms of all
evolutions.

Basically if there are multiple forms, it would use "Normal Form" (or
similar) while if there is only one form it would use "" as the form
name.