Support new Pokemon Go version
As a first quick patch, it only modifies how it reads the appraisal. It
tries to find the position of the box containing the stat bars.

If the color match, it compares the bars (or more precise one pixel row)
every 200 ms. It stops after 5 tries or when there are no changes and
then interprets that into an IV combination.

The checkboxes are necessary to say that a certain value is not actually
valid, for example before it has read the screen in or when the user is
manually inputting the values.